Prepend topic when ?recursive (NUL delimited)
Instead of just sending the body, also send the full topic, so
subscribers can make decisions based on the original topic.

The topic is ended by a NUL character: \x00. This was chosen
because %00s in URIs are recognised as being problematic and
"should be rejected if the application is not expecting ... raw data"