Rewrite standalone router to remove dup'd code
The router implementation had diverged since the interface was different
depending on whether the router was running in standalone (separate
binary) or integrated (built into the hookbot server) mode.

This was fixed when recursive topics were modified to prepend their
messages with the topic, followed by a null byte.

It only affects the router in standalone mode, and not the one we are
using in production (which is builtin to hookbot).